chore: mark as old for old schema
