+++ web/lib/schema/global-link.old.ts
@@ -1,5 +1,5 @@
 import { co, CoMap, Encoders } from "jazz-tools"
-import { GlobalTopic } from "./global-topic"
+import { GlobalTopic } from "./global-topic.old"
 
 // GlobalLinkAiSummary is high quality title, description, summary of link (generated by AI)
 export class GlobalLinkAiSummary extends CoMap {
