+++ web/lib/schema/global-topic.ts
@@ -1,21 +1,19 @@
-import { co, CoList, CoMap } from "jazz-tools"
-
-// GlobalTopic is topic that is accessible from `learn-anything.xyz/name`
+import { co, CoList, CoMap, Group } from "jazz-tools"
+export class TopicConnection extends CoMap {
+	name = co.string
+}
+export class ListOfTopicConnections extends CoList.Of(co.ref(TopicConnection)) {}
 export class GlobalTopic extends CoMap {
-	name = co.string // name of topic, used for url `learn-anything.xyz/name` (unique)
-	prettyName = co.string // pretty name of topic, e.g. "React"
-	aiSummary = co.optional.ref(GlobalTopicAiSummary) // high quality summary of topic (generated by AI)
-	humanSummary = co.optional.string // human created summary of link (simplified, cleaner)
-	topicAiSummary = co.optional.string // summary of topic (generated by AI)
-	topicSummary = co.optional.string // summary of topic (generated by AI)
-	public = co.optional.boolean // if true, topic is accessible from global search
+	name = co.string
+	prettyName = co.string
+	connections = co.optional.ref(ListOfTopicConnections)
 }
 
-// GlobalTopicAiSummary is high quality summary of topic (generated by AI)
-export class GlobalTopicAiSummary extends CoMap {
-	topic = co.ref(GlobalTopic) // topic the summary is for
-	aiModelUsed = co.string // model used to generate summary, e.g. "gpt-4o"
-	summary = co.string // high quality summary of topic (generated by AI)
-}
+export class ListOfGlobalTopics extends CoList.Of(co.ref(GlobalTopic)) {}
 
-export class GlobalTopicLists extends CoList.Of(co.ref(GlobalTopic)) {}
+export class PublicGlobalGroupRoot extends CoMap {
+	topics = co.optional.ref(ListOfGlobalTopics)
+}
+export class PublicGlobalGroup extends Group {
+	root = co.ref(PublicGlobalGroupRoot)
+}
