+++ web/lib/schema/index.ts
@@ -11,7 +11,7 @@
 import { CoMap, co, Account, Profile } from "jazz-tools"
 import { PersonalPageLists } from "./personal-page"
 import { PersonalLinkLists } from "./personal-link"
-import { GlobalTopicLists } from "./global-topic"
+import { ListOfGlobalTopics } from "./global-topic"
 
 export class UserRoot extends CoMap {
 	name = co.string
@@ -25,9 +25,9 @@ export class UserRoot extends CoMap {
 	personalPages = co.ref(PersonalPageLists)
 
 	// not implemented yet
-	topicsWantToLearn = co.ref(GlobalTopicLists)
-	topicsLearning = co.ref(GlobalTopicLists)
-	topicsLearned = co.ref(GlobalTopicLists)
+	topicsWantToLearn = co.ref(ListOfGlobalTopics)
+	topicsLearning = co.ref(ListOfGlobalTopics)
+	topicsLearned = co.ref(ListOfGlobalTopics)
 }
 
 export class LaAccount extends Account {
@@ -54,9 +54,9 @@ export class LaAccount extends Account {
 					personalPages: PersonalPageLists.create([], { owner: this }),
 
 					// not implemented yet
-					topicsWantToLearn: GlobalTopicLists.create([], { owner: this }),
-					topicsLearning: GlobalTopicLists.create([], { owner: this }),
-					topicsLearned: GlobalTopicLists.create([], { owner: this })
+					topicsWantToLearn: ListOfGlobalTopics.create([], { owner: this }),
+					topicsLearning: ListOfGlobalTopics.create([], { owner: this }),
+					topicsLearned: ListOfGlobalTopics.create([], { owner: this })
 				},
 				{ owner: this }
 			)
@@ -64,7 +64,6 @@ export class LaAccount extends Account {
 	}
 }
 
-export * from "./global-link"
 export * from "./global-topic"
 export * from "./personal-link"
 export * from "./personal-page"
